% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ezviewr-package.R
\docType{package}
\name{ezviewr-package}
\alias{ezviewr}
\alias{ezviewr-package}
\title{ezviewr: View Tidy Data in Preferable Spreadsheet}
\description{
Timely load your tidy-formatted data from R into a preferable spreadsheet application window.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jooyoungseo/ezviewr}
  \item Report bugs at \url{https://github.com/jooyoungseo/ezviewr/issues}
}

}
\author{
\strong{Maintainer}: JooYoung Seo \email{jooyoung@psu.edu}

}
\keyword{internal}
